dynamic partition table prototype working

+++ resources/generate_config.sh
@@ -1,6 +1,13 @@
 #!/bin/bash
+
 cd ..
 git submodule update --init --recursive
-cd src/defines/templates/
-cp gifnoc-template.h ../config.h
-cp laitnedifnoc-template.h ../confidential.h
+if [ ! -f "src/defines/config.h" ] || [ ! -f "src/defines/confidential.h" ]; then
+    echo -e ''
+    echo "Creating config template"
+    cd src/defines/templates/
+    cp gifnoc-template.h ../config.h
+    cp laitnedifnoc-template.h ../confidential.h
+    cd ../../../
+fi
+cd resources/
